Updating settings now works flawlessy
Before: if the user had no Twitter login permissions and he updated his
TW integration settings, the system would have attempted to log the
user in. But due to an API call mistake and due to the fact that the
code after an header() call is still executed by PHP, the system would
have outputted an awkward error.

Now: a patch has been included in this commit to prevent any
after-header() code to be executed. In addition to this, has been fixed
the action checkups to load the API regularly. Updating the settings
now SHOULD hopefully work even if you haven't an active access token
which is obtained and used to update the settings in one single step.
Tested it and it works, but it would be better to hear other's opinions
and feedbacks.